+++ app/javascript/src/components/packageBlocks/styled.js
@@ -1,0 +1,26 @@
+import styled from '@emotion/styled'
+import tw from 'twin.macro'
+import { keyframes } from '@emotion/core'
+
+const spin = keyframes`
+  100% { 
+    transform: rotate(360deg); 
+  } 
+`
+
+export const LoaderWrapper = styled.div`
+  ${() => tw`flex justify-center items-center`}
+`
+
+export const Progress = styled.div`
+  animation: ${spin} 0.5s infinite linear;
+  border-top-color: white !important;
+  
+  ${() => tw`
+    ease-linear rounded-full border-2 
+    border-t-2 border-gray-900 h-4 w-4
+    mx-auto
+    my-2
+  `}
+  
+`
